Rename utils/are_all_samples_control.py to utils.py
